Add namespace, but preserve 5.2 compatibility
This is a very quick implementation of a very rough idea.

The basic goal is to have the Parsedown class under a Parsedown namespace
for all of those who really want to use the library with a namespace.
But in the same time to not break the PHP 5.2 compatibility by just
adding a `namespace` line to the top of the file.

---

The current implementation consists of two Parsedown classes in separate files.
Each class extends ParsedownCore where the real code is. One of them have a
namespace and the other does not.