Add contribution guidelines file
Add initial version of contributing.md.

GitHub would show a link to this file on the new issue page and
when opening a PR.
https://help.github.com/articles/how-do-i-set-up-guidelines-for-contributors

Remove .gitignore as there is nothing to ignore from this project.
GitHub recommendation is to have just project-specific paths ignored in
the .gitignore file of the repo.
Platform-specific files and editor-specific files should go in a global
gitignore file.
https://help.github.com/articles/ignoring-files#global-gitignore